+++ src/lib/serialization/schema.ts
@@ -243,11 +243,9 @@ export interface QueryType extends Type, S<M.QueryType, "type" | "queryType"> {}
 
 export interface ReferenceType
     extends Type,
-        S<
-            M.ReferenceType,
-            "type" | "name" | "typeArguments" | "qualifiedName" | "package"
-        > {
+        S<M.ReferenceType, "type" | "name" | "typeArguments" | "package"> {
     id?: number;
+    qualifiedName?: string;
 }
 
 export interface ReflectionType extends Type, S<M.ReflectionType, "type"> {