+++ src/test/converter2/validation/namespace.ts
@@ -1,0 +1,5 @@
+export type Foo = Bar.Baz;
+
+namespace Bar {
+    export type Baz = 123;
+}