+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 -   TypeDoc will no longer crash if a comment contains an empty `@example` tag, #1967.
 -   TypeDoc will now detect attempted inheritance from accessors and inherit from the getter or setter, #1968.
+-   `intentionallyNotExported` will now properly respect qualified names, #1972.
+-   Validation warnings caused by missing documentation will now be formatted like other warnings which reference a declaration.
 
 ### Features